==> 1_extract_laion_landscapes.ipynb <==
%pip install pyspark

from pyspark.sql import SparkSession
from pyspark.sql.functions import lit, rand
# You can open http://localhost:4040 to follow progress on the spark operations
spark = SparkSession.builder.config("spark.driver.memory", "16G") .master("local[16]").appName('spark-stats').getOrCreate() 

parquetFile = spark.read.parquet("../../laion2B-en/")

KEYWORDS = ['garden','park[s ]','landscape','sand[sy ]','grass[sy ]','snow[sy ]','wind[sy ]','mountain','valley','waterfall','forest','hill','valley','^sky', ' sky ','cliff',' ridge','mist[sy ]','cloud[sy ]',' lake','countryside','sea ','ocean','beach']

parquetFile.filter(parquetFile.TEXT.rlike('|'.join(KEYWORDS))&(parquetFile.HEIGHT>250)& (parquetFile.NSFW=='UNLIKELY')&(parquetFile.similarity >0.5)).first()
