==> 2_download_dataset_images.ipynb <==
!wget -c http://images.cocodataset.org/zips/train2017.zip
!unzip -o train2017.zip
!wget -c http://images.cocodataset.org/zips/val2017.zip
!unzip -o val2017.zip

import pandas as pd
df= pd.read_csv("/content/dalle-tiny/tools/train/validation.csv")
import os

for index, (file_name,caption) in df.iterrows():
  os.rename("/content/val2017/"+file_name,'/content/archive_val/'+file_name)

!mkdir archive_val

!zip -r archive_val.zip archive_val/
