==> 3_encoding_image_with_vqgan_all_mscoco.ipynb <==
%pip install  git+https://github.com/CompVis/taming-transformers.git
!git clone https://github.com/CompVis/taming-transformers.git && cd taming-transformers && python -m pip install -e .
%pip install "omegaconf>=2.0.0" "pytorch-lightning>=1.0.8" einops transformers

import os
os._exit(00)

!wget http://images.cocodataset.org/zips/val2017.zip
!unzip val2017.zip

!wget http://images.cocodataset.org/zips/train2017.zip
!unzip train2017.zip

!wget http://images.cocodataset.org/zips/test2017.zip
!unzip test2017.zip

from tqdm import tqdm
import s3fs
import os

S3_ENDPOINT_URL = "https://" + os.environ["AWS_S3_ENDPOINT"]
fs = s3fs.S3FileSystem(client_kwargs={'endpoint_url': S3_ENDPOINT_URL})
S3_BUCKET="cthiounn2"

def write_file_to_s3(bucket_name:str,dir_file:str,file_name:str,fs:s3fs.core.S3FileSystem):
    if bucket_name and file_name and fs:
        FILE_PATH_OUT_S3 = bucket_name + "/" + file_name
        with fs.open(FILE_PATH_OUT_S3, 'wb') as file_out , open(dir_file+file_name, 'rb') as file_in:
            file_out.write(file_in.read())

files=['config_vqgan_minidalle.yaml','model_vqgan_minidalle.ckpt']
for file in tqdm(files):
    with fs.open(f'{S3_BUCKET}/{file}', mode="rb") as file_in, open(file,"wb") as file_out:
            file_out.write(file_in.read())


#VQGAN ImageNet (f=16), 1024

# !curl -L "https://heibox.uni-heidelberg.de/d/8088892a516d4e3baf92/files/?p=%2Fckpts%2Flast.ckpt&dl=1" >"last.ckpt"
# !curl -L "https://heibox.uni-heidelberg.de/d/8088892a516d4e3baf92/files/?p=%2Fconfigs%2Fmodel.yaml&dl=1" >"model.yaml"


#boris/vqgan_f16_16384
# !curl -L "https://huggingface.co/boris/vqgan_f16_16384/raw/main/config.yaml" > "config_vqgan_minidalle.yaml"
# !curl -L "https://huggingface.co/boris/vqgan_f16_16384/resolve/main/model.ckpt" > "model_vqgan_minidalle.ckpt"

import sys
sys.path.append(".")

# also disable grad to save memory
import torch
torch.set_grad_enabled(False)

DEVICE = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")

from omegaconf import OmegaConf
import taming
from taming.models.vqgan import VQModel

config_path = "config_vqgan_minidalle.yaml"
# config_path = "model.yaml"

config = OmegaConf.load(config_path)
model=VQModel(**config.model.params).to(DEVICE)

ckpt_path = "model_vqgan_minidalle.ckpt"
# ckpt_path = "last.ckpt"

sd = torch.load(ckpt_path, map_location=DEVICE)["state_dict"]
model.load_state_dict(sd, strict=False)

import torch
import torchvision.transforms.functional as TF
from torch.utils.data import Dataset, DataLoader
from torchvision.datasets.folder import default_loader
# from torchvision.transforms import InterpolationMode
import numpy as np
from PIL import Image
import torchvision.transforms as T
import requests

def preprocess_vqgan(x):
  x = 2.*x - 1.
  return x

def custom_to_pil(x):
  x = x.detach().cpu()
  x = torch.clamp(x, -1., 1.)
  x = (x + 1.)/2.
  x = x.permute(1,2,0).numpy()
  x = (255*x).astype(np.uint8)
  x = Image.fromarray(x)
  if not x.mode == "RGB":
    x = x.convert("RGB")
  return x

def resize_image(image):
      s = min(image.size)
      r = 256 / s
      s = (round(r * image.size[1]), round(r * image.size[0]))
      image = image.convert('RGB')
      image = TF.resize(image, s, interpolation=Image.LANCZOS)
      image = TF.center_crop(image, output_size = 2 * [256])
      image = torch.unsqueeze(T.ToTensor()(image), 0)

      return preprocess_vqgan(image.to(DEVICE))


from tqdm import tqdm
import pandas as pd
import os


def create_encoding_file_parquet(base_dir_images):
    list_file=[]
    list_encoded=[]
    for file_name in tqdm(os.listdir(base_dir_images), desc='dirs') :
        _, _, [_, _, indices]=model.encode(resize_image(Image.open(base_dir_images+"/"+file_name)))
        list_file.append(file_name)
        list_encoded.append(indices.tolist())
    df= pd.DataFrame.from_dict(
                    {"file_name": list_file, "encoding": list_encoded}
                )
    df.to_parquet(base_dir_images+".parquet")


create_encoding_file_parquet("val2017")
write_file_to_s3(S3_BUCKET,'./','val2017.parquet',fs)

create_encoding_file_parquet("train2017")
write_file_to_s3(S3_BUCKET,'./','train2017.parquet',fs)

create_encoding_file_parquet("test2017")
write_file_to_s3(S3_BUCKET,'./','test2017.parquet',fs)
