==> 4_prepare_dataset.ipynb <==
%run "1_extract_mscoco_landscapes.ipynb"

%store -r df_val_pair
%store -r df_train_pair

import pandas as pd

df_val=pd.read_parquet("https://github.com/cthiounn/dalle-tiny/raw/main/archive_val.parquet")
df_val=df_val.reset_index().set_index("file_name")
dfdf_val=df_val.merge(df_val_pair, how="inner",left_index=True, right_index=True)
dfdf_val=dfdf_val.reset_index()
dfdf_val=dfdf_val.drop(columns=['file_name','index'])
dfdf_val=dfdf_val.set_index("caption")
dfdf_val.to_parquet("val_caption_encoding.parquet")

df_train=pd.read_parquet("https://github.com/cthiounn/dalle-tiny/raw/main/archive_train.parquet")
df_train=df_train.reset_index().set_index("file_name")
dfdf_train=df_train.merge(df_train_pair, how="inner",left_index=True, right_index=True)
dfdf_train=dfdf_train.reset_index()
dfdf_train=dfdf_train.drop(columns=['file_name','index'])
dfdf_train=dfdf_train.set_index("caption")
dfdf_train.to_parquet("train_caption_encoding.parquet")
